Final refactor of standard shader generator - shader code moved to chunks
