+++ src/index.js
@@ -229,7 +229,6 @@ export { LitShaderOptions } from './scene/shader-lib/programs/lit-shader-options
 export { ProgramLibrary } from './scene/shader-lib/program-library.js';
 export { shaderChunks } from './scene/shader-lib/chunks/chunks.js';
 export { shaderChunksLightmapper } from './scene/shader-lib/chunks/chunks-lightmapper.js';
-export { ChunkBuilder } from './scene/shader-lib/chunk-builder.js';     // used by shed
 export { ChunkUtils } from './scene/shader-lib/chunk-utils.js';
 
 // SCENE / SKY
