+++ src/scene/shader-lib/chunks/chunks.js
@@ -175,6 +175,8 @@ import specularityFactorPS from './standard/frag/specularityFactor.js';
 import spotPS from './lit/frag/spot.js';
 import startNineSlicedPS from './lit/frag/startNineSliced.js';
 import startNineSlicedTiledPS from './lit/frag/startNineSlicedTiled.js';
+import stdDeclarationPS from './standard/frag/stdDeclaration.js';
+import stdFrontEndPS from './standard/frag/stdFrontEnd.js';
 import tangentBinormalVS from './lit/vert/tangentBinormal.js';
 import TBNPS from './lit/frag/TBN.js';
 import thicknessPS from './standard/frag/thickness.js';
@@ -384,6 +386,8 @@ const shaderChunks = {
     spotPS,
     startNineSlicedPS,
     startNineSlicedTiledPS,
+    stdDeclarationPS,
+    stdFrontEndPS,
     tangentBinormalVS,
     TBNPS,
     thicknessPS,
